feat: configure webpack for client app bundling & add babel config file
