+++ babel.config.cjs
@@ -1,4 +1,4 @@
-export default {
+module.exports = {
   presets: [
     ['@babel/preset-env', { useBuiltIns: 'usage' }],
     '@babel/preset-react',
