+++ client/index.jsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './components/App.jsx'
 import './assets/index.css'
