--- webpack.config.js
@@ -1,27 +1,0 @@
-import path from 'path'
-import HtmlWebpackPlugin from 'html-webpack-plugin'
-
-export default {
-  entry: './index.js',
-  mode: 'development',
-  plugins: [
-    new HtmlWebpackPlugin()
-  ],
-  output: {
-    filename: '[name].bundle.js',
-    path: path.resolve('dist')
-  },
-  module: {
-    rules: [
-      {
-        test: /\.css$/i,
-        use: ['style-loader', 'css-loader']
-      }
-    ]
-  },
-  devServer: {
-    static: './dist/index.html',
-    compress: true,
-    port: 3001
-  }
-}
